вынес форму в отдельный компонент
