+++ packages/client/src/fuature/login/components/AuthForm/AuthForm.module.scss
@@ -1,11 +1,11 @@
-.containerLogin {
+.containerForm {
 	width: 340px;
 	position: absolute;
 	top: 50%;
 	left: 50%;
 	border-radius: 20px;
 	background: var(--color-violet-darken);
-	margin: -228px 0 0 -170px;
+	margin: -151px 0 0 -170px;
 	padding: 35px 30px;
 	z-index: 9999;
 	display: flex;
@@ -14,6 +14,19 @@
 	align-items: center;
 	border-radius: 12px;
 
+	opacity: 0;
+	animation: ani 1s forwards;
+
+	@keyframes ani {
+		0% {
+			opacity: 0;
+		}
+
+		100% {
+			opacity: 1;
+		}
+	}
+
 	& h1 {
 		width: 100%;
 		color: var(--color-white);
@@ -21,13 +34,13 @@
 	}
 }
 
-.loginForm {
+.form {
 	display: flex;
 	flex-wrap: wrap;
 	background-color: var(--color-violet-darken);
 }
 
-.footerLogin {
+.footerForm {
 	position: absolute;
 	bottom: -50px;
 	color: var(--color-white);
