+++ packages/client/src/fuature/login/components/AuthForm/index.ts
@@ -1,0 +1,1 @@
+export * from './AuthForm';
