+++ packages/client/src/fuature/login/components/LoginForm/LoginForm.module.scss
@@ -1,39 +1,3 @@
 .containerLogin {
-	width: 340px;
-	position: absolute;
-	top: 50%;
-	left: 50%;
-	border-radius: 20px;
-	background: var(--color-violet-darken);
 	margin: -151px 0 0 -170px;
-	padding: 35px 30px;
-	z-index: 9999;
-	display: flex;
-	justify-content: center;
-	flex-direction: column;
-	align-items: center;
-	border-radius: 12px;
-
-	& h1 {
-		width: 100%;
-		color: var(--color-white);
-		font-size: 26px;
-	}
-}
-
-.loginForm {
-	display: flex;
-	flex-wrap: wrap;
-	background-color: var(--color-violet-darken);
-}
-
-.footerLogin {
-	position: absolute;
-	bottom: -50px;
-	color: var(--color-white);
-
-	& a {
-		color: #6781F7;
-		text-decoration: none;
-	}
 }
