+++ packages/client/src/fuature/login/components/LoginForm/LoginForm.tsx
@@ -1,14 +1,14 @@
 import { BodyNormal } from '@components/design-system/Fonts';
-
 import { Button } from '@components/design-system';
 import { Link } from 'react-router-dom';
 import styles from './LoginForm.module.scss';
 import { FormInput } from '@components/specific/FormInput/FormInput';
 import { useForm } from 'react-hook-form';
 import { RouteNames } from '@routes/routeNames';
-import { useState } from 'react';
+import { useState, useRef } from 'react';
 import { HidePassSVG } from '@components/design-system/SVG/HidePassSVG';
 import { ShowPassSVG } from '@components/design-system/SVG/ShowPassSVG';
+import { AuthForm } from '../AuthForm';
 
 type LoginT = {};
 
@@ -26,6 +26,7 @@ export const LoginForm: React.FC<LoginT> = () => {
 
   const {
     control,
+    watch,
     handleSubmit,
     formState: { errors },
   } = useForm<LoginFormT>({
@@ -35,9 +36,14 @@ export const LoginForm: React.FC<LoginT> = () => {
     },
   });
 
+  const password = useRef({});
+  password.current = watch('password', '');
+
   const toggleShowPassword = () => setIsPasswordShow((prev) => !prev);
 
   const showOrHidddenIcon = () => {
+    if (!watch().password.length) return;
+
     if (isPasswordShow) {
       return <HidePassSVG onClick={toggleShowPassword} />;
     } else {
@@ -47,43 +53,47 @@ export const LoginForm: React.FC<LoginT> = () => {
 
   const onSubmit = () => {};
 
+  const footer = () => {
+    return (
+      <BodyNormal weight={'normal'}>
+        Нет аккаунта? <Link to={RouteNames.REGISTRATION}>Зарегистрироваться</Link>
+      </BodyNormal>
+    );
+  };
+
   return (
-    <div className={styles.containerLogin}>
-      <h1>Вход</h1>
-      <form className={styles.loginForm} onSubmit={handleSubmit(onSubmit)}>
-        <FormInput
-          name="login"
-          label="Введите электронную почту"
-          control={control}
-          rules={{
-            required: 'Это поле обязательно',
-            pattern: {
-              value: /\S+@\S+\.\S+/,
-              message: 'Некорректный адрес электронной почты',
-            },
-          }}
-          style={{ marginTop: '22px' }}
-        />
-        <FormInput
-          name="password"
-          label="Введите пароль"
-          type={isPasswordShow ? 'text' : 'password'}
-          control={control}
-          rules={{
-            required: 'Это поле обязательно',
-          }}
-          rightAddon={showOrHidddenIcon()}
-          style={{ marginTop: '22px' }}
-        />
-        <Button color={'pink'} style={{ marginTop: '22px' }} type="submit">
-          <BodyNormal weight={'normal'}>Войти</BodyNormal>
-        </Button>
-      </form>
-      <div className={styles.footerLogin}>
-        <BodyNormal weight={'normal'}>
-          Нет аккаунта? <Link to={RouteNames.REGISTRATION}>Зарегистрироваться</Link>
-        </BodyNormal>
-      </div>
-    </div>
+    <AuthForm title="Вход" onSubmit={handleSubmit(onSubmit)} footer={footer()} className={styles.containerLogin}>
+      <FormInput
+        name="login"
+        label="Введите электронную почту"
+        control={control}
+        rules={{
+          required: 'Это поле обязательно',
+          pattern: {
+            value: /\S+@\S+\.\S+/,
+            message: 'Некорректный адрес электронной почты',
+          },
+        }}
+        style={{ marginTop: '22px' }}
+      />
+      <FormInput
+        name="password"
+        label="Введите пароль"
+        type={isPasswordShow ? 'text' : 'password'}
+        control={control}
+        rules={{
+          required: 'Это поле обязательно',
+          pattern: {
+            value: /^(?=.*[a-zA-Z])(?=.*\d)[A-Za-z\d@$!%*?&]{8,}$/,
+            message: 'min 8 символов, min 1 цифра и 1 загл. буква',
+          },
+        }}
+        rightAddon={showOrHidddenIcon()}
+        style={{ marginTop: '22px' }}
+      />
+      <Button color={'pink'} style={{ marginTop: '22px' }} type="submit">
+        <BodyNormal weight={'normal'}>Войти</BodyNormal>
+      </Button>
+    </AuthForm>
   );
 };
