+++ packages/client/src/fuature/registration/components/RegistrationForm/RegistrationForm.module.scss
@@ -1,0 +1,3 @@
+.containerRegistration {
+	margin: -300px 0 0 -170px;
+}
