+++ packages/client/src/fuature/registration/components/RegistrationForm/index.ts
@@ -1,0 +1,1 @@
+export * from '.';
