--- packages/client/src/fuature/registration/components/index.ts
@@ -1,1 +1,0 @@
-export * from './RegistrationForm';
